+++ json/json-module/src/main/java/ru/tinkoff/kora/json/module/http/JsonHttpBodyOutput.java
@@ -26,7 +26,7 @@ public final class JsonHttpBodyOutput<T> implements HttpBodyOutput {
     }
 
     @Override
-    public int contentLength() {
+    public long contentLength() {
         return -1;
     }
